User can now change the layout predefinned layout in db and moved databases sub-folder to reorganize
