+++ cogs/archiveuse/smartpoint.py
@@ -2,7 +2,7 @@ import traceback
 import  discord
 from libs.config import Config
 
-from libs.databases.user import User
+from libs.databases.user.user import User
 from libs.exception.handler import Handler
 from libs.log import Log, LogType
