+++ cogs/barman/root_commands.py
@@ -2,10 +2,10 @@ import traceback
 import discord
 import requests
 from libs.config import Config
-from libs.databases.user import User
-from libs.databases.users import Users
-from libs.databases.wallpaper import Wallpaper
-from libs.databases.wallpapers import Wallpapers
+from libs.databases.user.user import User
+from libs.databases.user.users import Users
+from libs.databases.wallpaper.wallpaper import Wallpaper
+from libs.databases.wallpaper.wallpapers import Wallpapers
 from libs.exception.color.color_not_correct_exception import ColorNotCorrectException
 from libs.exception.handler import Handler
 from libs.exception.wallpaper.wallpaper_already_exist_exception import WallpaperAlreadyExistException
