+++ data/databases/sqliteDB.sql
@@ -47,6 +47,7 @@ CREATE TABLE IF NOT EXISTS usersHaveBadge (
 
 CREATE TABLE IF NOT EXISTS profilesLayout (
   id INTEGER PRIMARY KEY AUTOINCREMENT,
+  name TEXT NOT NULL UNIQUE,
   profilPictureX INTEGER NOT NULL,
   profilPictureY INTEGER NOT NULL,
   nameX INTEGER NOT NULL,
@@ -61,11 +62,8 @@ CREATE TABLE IF NOT EXISTS profilesLayout (
   levelBarY INTEGER NOT NULL
 );
 
-INSERT INTO profilesLayout (id, profilPictureX, profilPictureY, nameX, nameY, userNameX, userNameY, levelX, levelY, badgeX, badgeY, levelBarX, levelBarY) VALUES 
-(1, 0, 0, 150, 20, 150, 65, 250, 224, 150, 90, 0, 254);
-
--- INSERT INTO profilesLayout (id, profilPictureX, profilPictureY, nameX, nameY, userNameX, userNameY, levelX, levelY, badgeX, badgeY, levelBarX, levelBarY) VALUES 
--- (2, 186, 35, 186, 155, 190, 195, 250, 224, 150, 5, 0, 254);
-
 INSERT INTO wallpapers (id, name, url, level, price) VALUES 
 (1, 'default', 'https://shkermit.ch/~ethann/compHydromel/wallpapers/default.png', 0, 0);
+
+INSERT INTO profilesLayout (id, name, profilPictureX, profilPictureY, nameX, nameY, userNameX, userNameY, levelX, levelY, badgeX, badgeY, levelBarX, levelBarY) VALUES 
+(1, "default", 0, 0, 150, 20, 150, 65, 250, 224, 150, 90, 0, 254);
