+++ libs/databases/adaptaters/database_access_implement.py
@@ -1,16 +1,16 @@
-from libs.databases.database_access_implement import DatabaseAccessImplement, ProfileColoredPart
-from libs.databases.sqlite.sqlite import Sqlite
-from typing import Any
+from enum import Enum
 
-class SqliteAccess(DatabaseAccessImplement):
-    __sqliteDB : Sqlite
-    
+class ProfileColoredPart(Enum):
+    BarColor = "barColor"
+    NameColor = "nameColor"
+
+class DatabaseAccessImplement:
+    """This class is designed to be the interface of the database access.
+    """
     def __init__(self) -> None:
-        """This method is designed to initialize the database. Used to initiate the databases and create db if not exist.
+        """This method is designed to initialize the class. Use it to initialize the database connection.
         """
-        self.__sqliteDB = Sqlite("database.db")
-        if len(self.__sqliteDB.select("SELECT name FROM sqlite_master WHERE type='table' AND name='users';")) == 0:
-            self.__create_db()
+        pass
 
     # Public 
 
@@ -23,7 +23,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             int: The user level number.
         """
-        return self.__select_user(discord_id, "level")[0][0]
+        pass
 
     def get_user_point(self, discord_id: str) -> int:
         """This method is designed to get a user number of point.
@@ -34,7 +34,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             int: The user number of point.
         """
-        return self.__select_user(discord_id, "point")[0][0]
+        pass
     
     def get_smartpoint(self, discord_id: str) -> int:
         """This method is designed to get a user number of smartpoint.
@@ -45,7 +45,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             int: The user number of smartpoint.
         """
-        return self.__select_user(discord_id, "smartpoint")[0][0]
+        pass
     
     def get_number_of_buy(self, discord_id: str) -> int:
         """This method is designed to get a user number of buy.
@@ -56,7 +56,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             int: The user number of buy.
         """
-        return self.__select_user(discord_id, "numberOfBuy")[0][0]
+        pass
     
     def add_user_point(self, discord_id: str, point: int = 1) -> None:
         """This method is designed to add point to a user.
@@ -65,7 +65,7 @@ class SqliteAccess(DatabaseAccessImplement):
             discord_id (str): Discord user id as a string.
             point (int, optional): The number of point to add. Defaults to 1.
         """
-        self.__sqliteDB.modify("UPDATE users SET point = point + " + str(point) + " WHERE discordId = '" + discord_id + "'")
+        pass
 
     def add_user_level(self, discord_id: str, level = 1) -> None:
         """This method is designed to add level to a user.
@@ -74,7 +74,7 @@ class SqliteAccess(DatabaseAccessImplement):
             discord_id (str): Discord user id as a string.
             level (int, optional): The number of level to add. Defaults to 1.
         """
-        self.__sqliteDB.modify("UPDATE users SET level = level + " + str(level) + " WHERE discordId = '" + discord_id + "'")
+        pass
 
     def add_smartpoint(self, discord_id, amount=1) -> None:
         """This method is designed to add smartpoint to a user.
@@ -83,8 +83,8 @@ class SqliteAccess(DatabaseAccessImplement):
             discord_id (_type_): Discord user id as a string.
             amount (int, optional): The number of smartpoint to add. Defaults to 1.
         """
-        self.__sqliteDB.modify("UPDATE users SET smartpoint =  smartpoint + " + str(amount) + " WHERE discordId = '" + discord_id + "'")
-    
+        pass
+
     def remove_smartpoint(self, discord_id, amount=1) -> None:
         """This method is designed to remove smartpoint to a user.
 
@@ -92,17 +92,15 @@ class SqliteAccess(DatabaseAccessImplement):
             discord_id (_type_): Discord user id as a string.
             amount (int, optional): The number of smartpoint to remove. Defaults to 1.
         """
-        self.__sqliteDB.modify("UPDATE users SET smartpoint = smartpoint - " + str(amount) + " WHERE discordId = '" + discord_id + "'")
-    
+        pass
+
     def add_user_if_not_exist(self, discord_id: str) -> None:
         """This method is designed to add a user if he doesn't exist.
         
         Args:
             discord_id (str): Discord user id as a string.
         """
-        if not self.__sqliteDB.select("SELECT discordId FROM users WHERE discordId = '" + discord_id + "';"):
-            self.__sqliteDB.modify("INSERT INTO users(discordId) VALUES ('" + discord_id + "')")
-            self.add_posseded_wallpaper(discord_id, "default")
+        pass
 
     def reset_point(self, discord_id: str) -> None:
         """This method is designed to reset a user point to 0.
@@ -110,15 +108,15 @@ class SqliteAccess(DatabaseAccessImplement):
         Args:
             discord_id (str): Discord user id as a string.
         """
-        self.__sqliteDB.modify("UPDATE users SET point = 0 WHERE discordId = '" + discord_id + "';")
-
+        pass
+    
     def reset_level(self, discord_id: str) -> None:
         """This method is designed to reset a user level to 1.
 
         Args:
             discord_id (str): Discord user id as a string.
         """
-        self.__sqliteDB.modify("UPDATE users SET level = 1 WHERE discordId = '" + discord_id + "';")
+        pass
 
     def get_user_profile_custom_color(self, discord_id: str, profile_colored_part: ProfileColoredPart) -> str:
         """This method is designed to get a user profile custom color.
@@ -130,7 +128,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             str: The user custom bar color as Hex RGB (example: 00ff00, ff00ffaf, etc..).
         """
-        return self.__select_user(discord_id, profile_colored_part.value)[0][0]
+        pass
 
     def get_if_user_is_root(self, discord_id: str) -> bool:
         """This method is designed to get if a user is root.
@@ -141,8 +139,8 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             bool: True if the user is root, False if not.
         """
-        return bool(self.__select_user(discord_id, "isRoot")[0][0])
-
+        pass
+    
     def set_user_root(self, discord_id: str, is_root: bool) -> None:
         """This method is designed to set a user root.
 
@@ -150,7 +148,7 @@ class SqliteAccess(DatabaseAccessImplement):
             discord_id (str): Discord user id as a string.
             is_root (bool): True if the user is root, False if not.
         """
-        self.__sqliteDB.modify("UPDATE users SET isRoot = " + str(int(is_root)) + " WHERE discordId = '" + discord_id + "';")
+        pass
 
     def get_user_current_wallpaper(self, discord_id: str) -> str:
         """This method is designed to get a user current wallpaper name.
@@ -161,11 +159,8 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             str: The user current wallpaper name. Return "default" if the user has a deleted wallpaper put on.
         """
-        query_result = self.__sqliteDB.select("SELECT wallpapers.name FROM users INNER JOIN wallpapers ON users.wallpapersId = wallpapers.id WHERE discordId = '" + discord_id + "';")
-        if len(query_result) > 0:
-            return query_result[0][0]
-        return "default"
-        
+        pass
+
     def get_list_posseded_wallpapers(self, discord_id: str) -> list[str]:
         """This method is designed to get a user posseded wallpapers list.
 
@@ -175,10 +170,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             list[str]: The user posseded wallpapers name list (example: ['default', 'wallpaper_name']).
         """
-        return self.__reorder_list(self.__sqliteDB.select("SELECT wallpapers.name FROM users \
-                                      INNER JOIN usersBuyWallpapers ON usersBuyWallpapers.usersId = users.id \
-                                      INNER JOIN wallpapers ON usersBuyWallpapers.wallpapersId = wallpapers.id \
-                                      WHERE discordId = '" + discord_id + "';"))
+        pass
 
     def change_user_current_wallpaper(self, discord_id: str, wallpaper_name: str) -> None:
         """This method is designed to change a user current wallpaper.
@@ -187,22 +179,16 @@ class SqliteAccess(DatabaseAccessImplement):
             discord_id (str): Discord user id as a string.
             wallpaper_name (str): The new wallpaper name.
         """
-        return self.__sqliteDB.modify("WITH updated_wallpapers AS (\
-                                        SELECT id\
-                                        FROM wallpapers\
-                                        WHERE name = '" + wallpaper_name + "')\
-                                    UPDATE users\
-                                    SET wallpapersId = (SELECT id FROM updated_wallpapers)\
-                                    WHERE discordId = '" + discord_id + "';")
-    
+        pass
+
     def get_all_wallpapers(self) -> list[str]:
         """This method is designed to get all wallpapers name.
 
         Returns:
             list[str]: All wallpapers name (example: ['default', 'wallpaper_name']).
         """         
-        return self.__reorder_list(self.__sqliteDB.select("SELECT name FROM wallpapers"))
-    
+        pass
+
     def is_wallpaper_exist(self, wallpaper_name: str) -> bool:
         """This method is designed to check if a wallpaper exist.
 
@@ -212,8 +198,8 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             bool: True if the wallpaper exist, False if not.
         """
-        return len(self.__sqliteDB.select("SELECT name FROM wallpapers WHERE name = '" + wallpaper_name + "'")) > 0
-    
+        pass
+
     def get_wallpaper_price(self, wallpaper_name: str) -> int:
         """This method is designed to get a wallpaper price.
 
@@ -223,8 +209,8 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             int: The wallpaper price.
         """
-        return self.__select_wallpaper(wallpaper_name, "price")[0][0]
-    
+        pass
+
     def get_wallpaper_url(self, wallpaper_name: str) -> str:
         """This method is designed to get a wallpaper url.
 
@@ -234,7 +220,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             str: The wallpaper url (example: https://example.com/img.png).
         """
-        return self.__select_wallpaper(wallpaper_name, "url")[0][0]
+        pass
     
     def get_wallpaper_level(self, wallpaper_name: str) -> str:
         """This method is designed to get a wallpaper level.
@@ -245,7 +231,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             str: The wallpaper level.
         """
-        return self.__select_wallpaper(wallpaper_name, "level")[0][0]
+        pass
 
     def change_user_profile_custom_color(self, discord_id: str, profile_colored_part: ProfileColoredPart, color: str) -> None:
         """This method is designed to change a user profile custom color.
@@ -255,29 +241,26 @@ class SqliteAccess(DatabaseAccessImplement):
             profile_colored_part (ProfileColoredPart): The profile colored part (NameColor or BarColor).
             color (str): The new color as Hex RGB (example: 00ff00, ff00ffaf, etc..).
         """
-        self.__sqliteDB.modify("UPDATE users SET " + profile_colored_part.value + " = '" + color + "' WHERE discordId = '" + discord_id + "';")
-
-    def get_users_badge_list(self, discord_id: str) -> list[str]:
+        pass
+    
+    def get_users_badge_list(self, discord_id: str) -> list[list[str]]:
         """This method is designed to get a user badge list.
 
         Args:
             discord_id (str): Discord user id as a string.
 
         Returns:
-            list[str]: The user badge list (example: ['badge_name', 'badge_name']).
+            list[list[str]]: The user badge list (example: [['badge_name', 'https://badge_url'], ['badge_name', 'https://badge_url']]).
         """
-        return self.__reorder_list(self.__sqliteDB.select("SELECT badges.name FROM users \
-                                      INNER JOIN usersHaveBadge ON usersHaveBadge.usersId = users.id \
-                                      INNER JOIN badges ON usersHaveBadge.badgesId = badges.id \
-                                      WHERE users.discordId = '" + discord_id + "';"))
-
+        pass
+    
     def get_top_users(self) -> list[str]:
         """This method is designed to get the top users.
 
         Returns:
             list[str]: The top users list (example: ['discord_id', 'discord_id']).
         """
-        return self.__reorder_list(self.__sqliteDB.select("SELECT discordId FROM users ORDER BY level DESC, point DESC LIMIT 10"))
+        pass
     
     def get_root_users(self) -> list[str]:
         """This method is designed to get the root users.
@@ -285,8 +268,8 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             list[str]: The root users list (example: ['discord_id', 'discord_id']).
         """
-        return self.__reorder_list(self.__sqliteDB.select("SELECT discordId FROM users WHERE isRoot = 1"))
-    
+        pass
+
     def add_posseded_wallpaper(self, discordId: str, wallpaper_name: str) -> None:
         """This method is designed to add a wallpaper to a user's possession.
 
@@ -294,13 +277,8 @@ class SqliteAccess(DatabaseAccessImplement):
             discordId (str): Discord user id as a string.
             wallpaper_name (str): The wallpaper name.
         """
-        self.__sqliteDB.modify("INSERT INTO usersBuyWallpapers (usersId, wallpapersId) \
-                                SELECT u.id AS userId, w.id AS wallpaperId \
-                                FROM users u \
-                                CROSS JOIN wallpapers w \
-                                WHERE u.discordId = '" + discordId + "' \
-                                AND w.name = '" + wallpaper_name + "';")
-        
+        pass
+    
     def is_badge_exist(self, badge_name: str) -> bool:
         """This method is designed to check if a badge exist.
 
@@ -310,7 +288,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             bool: True if the badge exist, False if not.
         """
-        return len(self.__sqliteDB.select("SELECT name FROM badges WHERE name = '" + badge_name + "'")) > 0
+        pass
     
     def get_badge_url(self, badge_name: str) -> str:
         """This method is designed to get a badge url.
@@ -321,7 +299,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Returns:
             str: The badge url (example: https://example.com/img.png).
         """
-        return self.__sqliteDB.select("SELECT url FROM badges WHERE name = '" + badge_name + "'")[0][0]
+        pass
     
     def increase_number_of_buy(self, discord_id: str) -> None:
         """This method is designed to increase the number of buy of a user.
@@ -329,7 +307,7 @@ class SqliteAccess(DatabaseAccessImplement):
         Args:
             discord_id (str): Discord user id as a string.
         """
-        self.__sqliteDB.modify("UPDATE users SET numberOfBuy = numberOfBuy + 1 WHERE discordId = '" + discord_id + "';")
+        pass
 
     def reset_number_of_buy(self, discord_id: str) -> None:
         """This method is designed to reset the number of buy of a user.
@@ -337,45 +315,57 @@ class SqliteAccess(DatabaseAccessImplement):
         Args:
             discord_id (str): Discord user id as a string.
         """
-        self.__sqliteDB.modify("UPDATE users SET numberOfBuy = 0 WHERE discordId = '" + discord_id + "';")
-
+        pass
+    
     def get_user_profile_layout(self, discord_id: str) -> dict[str,dict[str, int]]:
-        """This method is designed to get users profiles layout.
+        """This method is designed to get user profile layout.
 
         Args:
             discord_id (str): Discord user id as a string.
-
+        
         Returns:
             dict[dict[str, int]]: The users profiles coords list (example: {"profilPicture": {"x": 0, "y": 0}, "name": ...}).
         """
-        raw = self.__sqliteDB.select("SELECT profilPictureX, profilPictureY, nameX, nameY, userNameX, userNameY, levelX, levelY, badgeX, badgeY, levelBarX, levelBarY FROM profilesLayout INNER JOIN users ON users.profilesLayoutId = profilesLayout.id WHERE discordId = '" + discord_id + "';")[0]
-        
-        return {
-            "profilPicture": {
-                "x": raw[0],
-                "y": raw[1]
-            },
-            "name": {
-                "x": raw[2],
-                "y": raw[3]
-            },
-            "userName": {
-                "x": raw[4],
-                "y": raw[5]
-            },
-            "level": {
-                "x": raw[6],
-                "y": raw[7]
-            },
-            "badge": {
-                "x": raw[8],
-                "y": raw[9]
-            },
-            "levelBar": {
-                "x": raw[10],
-                "y": raw[11]
-            }
-        }
+        pass
+    
+    def change_user_profile_layout(self, discord_id: str, layout_name: str) -> None:
+        """This method is designed to set user profile layout.
+
+        Args:
+            discord_id (str): Discord user id as a string.
+            layout_name (str): The layout name.
+        """
+        pass
+    
+    def is_profile_layout_exist(self, layout_name: str) -> bool:
+        """This method is designed to check if a profile layout exist.
+
+        Args:
+            layout_name (str): The layout name.
+
+        Returns:
+            bool: True if the user profile layout exist, False if not.
+        """
+        pass
+    
+    def get_all_profile_layouts(self) -> list[str]:
+        """This method is designed to list all user profile layout.
+
+        Returns:
+            list[str]: The list of all user profile layout name.
+        """
+        pass
+    
+    def get_profile_layout(self, layout_name: str) -> dict[str,dict[str, int]]:
+        """This method is designed to get a profile layout.
+
+        Args:
+            layout_name (str): The layout name.
+
+        Returns:
+            dict[dict[str, int]]: The profile layout coords list (example: {"profilPicture": {"x": 0, "y": 0}, "name": ...}).
+        """
+        pass
     
     def add_wallpaper(self, wallpaper_name: str, url: str, price: int, level: int) -> None:
         """This method is designed to add a wallpaper to the database.
@@ -386,7 +376,7 @@ class SqliteAccess(DatabaseAccessImplement):
             price (int): The price of the wallpaper.
             level (int): The level of the wallpaper.
         """
-        self.__sqliteDB.modify("INSERT INTO wallpapers(name, url, price, level) VALUES ('" + wallpaper_name + "', '" + url + "', " + str(price) + ", " + str(level) + ");")
+        pass
     
     def remove_wallpaper(self, wallpaper_name: str) -> None:
         """This method is designed to remove a wallpaper from the database.
@@ -394,54 +384,6 @@ class SqliteAccess(DatabaseAccessImplement):
         Args:
             wallpaper_name (str): The wallpaper name.
         """
-        self.__sqliteDB.modify("DELETE FROM wallpapers WHERE name = '" + wallpaper_name + "';")
+        pass
 
     # Public
-
-    # Private
-
-    def __select_user(self, discord_id: str, selection) -> list[list[Any]]:
-        """This method is designed to select a user's data.
-
-        Args:
-            discord_id (str): Discord user id as a string.
-            selection (_type_): The data to select. (example: level, point, discordId, etc.. as string with , between each selection)
-
-        Returns:
-            list[list[Any]]: The user's data.
-        """
-        return self.__sqliteDB.select("SELECT " + selection + " FROM users WHERE discordId = '" + discord_id + "'")
-    
-    def __select_wallpaper(self, wallpaper_name: str, selection) -> list[list[Any]]:
-        """This method is designed to select a wallpaper's data.
-
-        Args:
-            wallpaper_name (str): The wallpaper name.
-            selection (_type_): The data to select. (example: level, price, url, etc.. as string with , between each selection)
-
-        Returns:
-            list[list[Any]]: The wallpaper's data.
-        """
-        return self.__sqliteDB.select("SELECT " + selection + " FROM wallpapers WHERE name = '" + wallpaper_name + "'")
-    
-    def __create_db(self) -> None:
-        """This method is designed to create the database.
-        """
-        with open("data/databases/sqliteDB.sql", "r") as script:
-            for cmd in script.read().split(";"):
-                self.__sqliteDB.modify(cmd)
-    
-    def __reorder_list(self, list_to_reorder: list[list[Any]]) -> list[any]:
-        """This method is designed to reorder a list.
-
-        Args:
-            list_to_reorder (list[list[Any]]): The list to reorder.
-
-        Returns:
-            list[any]: The reordered list.
-        """
-        list_reorder = []
-        for item in list_to_reorder:
-            list_reorder.append(item[0])
-        return list_reorder
-    # Private
