+++ libs/databases/adaptaters/sqlite/sqlite.py
