+++ libs/databases/badge/badge.py
@@ -1,5 +1,5 @@
-from libs.databases.database_access_implement import DatabaseAccessImplement
-from libs.databases.sqlite.sqlite_access import SqliteAccess
+from libs.databases.adaptaters.database_access_implement import DatabaseAccessImplement
+from libs.databases.adaptaters.sqlite.sqlite_access import SqliteAccess
 from libs.exception.badge.badge_not_exist_exception import BadgeNotExistException
