+++ libs/databases/badge/badges.py
@@ -1,6 +1,6 @@
-from libs.databases.badge import Badge
-from libs.databases.database_access_implement import DatabaseAccessImplement
-from libs.databases.sqlite.sqlite_access import SqliteAccess
+from libs.databases.badge.badge import Badge
+from libs.databases.adaptaters.database_access_implement import DatabaseAccessImplement
+from libs.databases.adaptaters.sqlite.sqlite_access import SqliteAccess
 
 
 class Badges:
