+++ libs/databases/wallpaper/wallpapers.py
@@ -1,6 +1,6 @@
-from libs.databases.database_access_implement import DatabaseAccessImplement
-from libs.databases.sqlite.sqlite_access import SqliteAccess
-from libs.databases.wallpaper import Wallpaper
+from libs.databases.adaptaters.database_access_implement import DatabaseAccessImplement
+from libs.databases.adaptaters.sqlite.sqlite_access import SqliteAccess
+from libs.databases.wallpaper.wallpaper import Wallpaper
 from libs.exception.wallpaper.wallpaper_already_exist_exception import WallpaperAlreadyExistException
 from libs.exception.wallpaper.wallpaper_not_exist_exception import WallpaperNotExistException
