+++ libs/exception/handler.py
@@ -59,6 +59,8 @@ class Handler:
                     return self.__response_exception["nothing_left_in_queue"]
                 case "NoMusicPlaying":
                     return self.__response_exception["not_playing_music"]
+                case "ProfileLayoutNotExist":
+                    return self.__response_exception["profile_layout_not_exist"]
                 case _:
                     return self.__default(stacktrace)
         else:
