+++ libs/exception/profile_layout/profile_layout_not_exist.py
@@ -1,0 +1,4 @@
+from libs.exception.bot_exception import BotException
+
+class ProfileLayoutNotExist(BotException):
+    pass
